new changes were made

+++ applications/model/cdn-client-subflow.h
@@ -115,7 +115,7 @@ private:
    
   Ptr<CdnClient> m_parent; //!<parent connection.
 
-  uint32_t m_count; //!< Maximum number of packets the application will send
+  uint64_t m_count; //!< Maximum number of packets the application will send
   Time m_interval; //!< Packet inter-send time
   uint32_t m_size; //!< Size of the sent packet (including the SeqTsHeader)
   Ptr<Socket> m_socket; //!< Socket
